src/test/java/com/zcy/restaurant/service/impl/SellerServiceImplTest.java
package com.zcy.restaurant.service.impl;

import com.zcy.restaurant.entity.SellerInfo;
import com.zcy.restaurant.service.SellerService;
import lombok.extern.slf4j.Slf4j;
import org.junit.Assert;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.stereotype.Service;
import org.springframework.test.context.junit4.SpringRunner;

import javax.swing.*;

import static org.junit.Assert.*;

@RunWith(SpringRunner.class)
@SpringBootTest
@Slf4j
public class SellerServiceImplTest {

    private static final String openid = "abc";

    @Autowired
    private SellerServiceImpl sellerServiceImpl;

    @Test
    public void findSellerInfoByOpenid() {
        SellerInfo result = sellerServiceImpl.findSellerInfoByOpenid(openid);
        Assert.assertEquals(openid, result.getOpenid());
    }
}